A few improvement to different elements
Game objects can only be created through a scene, by adding a game object to it, which will return that object.
The nodes of the maze are now components attached to game objects who's parent is the maze object.
The render texture function now has a offset parameter to draw a texture with it's center on the position of the object.
A tank object now requires to be spawned on a maze spawnpoint.
Removed outdated gameobject hierarchy from the gameobject class.